🌟 split user and channel management. add condition for use clear comand (have permission to manage messages)
