+++ src/fablabot/cogs/user_management.py
@@ -1,4 +1,4 @@
-"""Channel management commands and permission utilities."""
+"""User management commands and permission utilities."""
 
 from __future__ import annotations
 
@@ -17,90 +17,6 @@ from fablabot.role_loader import can_assign_role, load_roles
 logger = logging.getLogger(__name__)
 
 CURRENT_TIME = datetime.now().strftime("%Y/%m/%d, %H:%M:%S")
-
-# define the different permissions [admin, invited, read only , blacklist]
-# TODO: replace with named tuple
-permissions = {
-    "send_messages": [True, True, False, False],
-    "read_messages": [True, True, True, False],
-    "manage_messages": [True, False, False, False],
-    "manage_channels": [True, False, False, False],
-    "manage_roles": [True, False, False, False],
-    "manage_permissions": [True, False, False, False],
-    "manage_emojis": [True, True, False, False],
-    "mention_everyone": [True, False, False, False],
-    "create_private_threads": [True, True, False, False],
-    "create_public_threads": [True, True, False, False],
-    "read_message_history": [True, True, True, False],
-    "add_reactions": [True, True, True, False],
-    "attach_files": [True, True, False, False],
-}
-overwrite = [discord.PermissionOverwrite() for _ in range(4)]
-
-for perm_name in permissions:
-    for level in range(len(overwrite)):
-        overwrite[level].update(**{perm_name: permissions[perm_name][level]})
-overwrite += [None]
-
-
-class ChannelManagement(app_commands.Group, name="channel", description="Gestion des salons"):
-    """Manages channel-related commands.
-
-    Args:
-        app_commands (app_commands.Group): The app_commands group.
-        name (str, optional): The name of the group. Defaults to "channel".
-        description (str, optional): The description of the group. Defaults to "Gestion des salons".
-    """
-
-    @app_commands.command(name="clear", description="Nettoie le salon actuel de ses derniers 100 messages.")
-    async def clear(self, interaction: discord.Interaction) -> None:
-        """Clears the current channel of its last 100 messages.
-
-        Args:
-            interaction: The interaction that triggered the command.
-        """
-        logger.info("%s clear %s:%s", CURRENT_TIME, interaction.user.name, interaction.user.id)
-        if isinstance(
-            interaction.channel,
-            discord.TextChannel | discord.channel.VocalGuildChannel | discord.Thread,
-        ):
-            await interaction.channel.purge(limit=100)
-
-    @app_commands.command(name="create", description="Crée un nouveau salon dans la catégorie spécifiée.")
-    @app_commands.describe(
-        channel="Le nom du salon à créer",
-        category="La catégorie dans laquelle créer le salon",
-    )
-    async def create(
-        self,
-        interaction: discord.Interaction,
-        channel: str,
-        category: discord.CategoryChannel,
-    ) -> None:
-        """Create a new channel in the passed category.
-
-        Args:
-            interaction (discord.Interaction): The interaction object.
-            channel (str): The name of the channel to create.
-            category (discord.CategoryChannel): The category to create the channel in.
-        """
-        assert interaction.guild is not None  # Satisfies type checker
-        assert isinstance(interaction.user, discord.Member)  # Satisfies type checker
-        logger.info(
-            "%s create %s:%s %s %s",
-            CURRENT_TIME,
-            interaction.user.name,
-            interaction.user.id,
-            channel,
-            category,
-        )
-        if channel not in [channel.name for channel in category.channels]:
-            new_channel = await interaction.guild.create_text_channel(channel, category=category)
-            await new_channel.set_permissions(interaction.user, overwrite=overwrite[0])
-
-            await interaction.response.send_message(f"Le salon {channel!r} a été créé dans {category.name} !")
-        else:
-            await interaction.response.send_message(f"Un salon {channel!r} existe déjà dans {category.name} !")
 
 
 class UserManagementGroup(app_commands.Group, name="user", description="Gestion des utilisateurs"):
@@ -322,10 +238,7 @@ class UserManagement(commands.Cog):
             bot: The bot instance.
         """
         self.bot = bot
-        for group in (
-            ChannelManagement(),
-            UserManagementGroup(),
-        ):
+        for group in (UserManagementGroup(),):
             self.bot.tree.add_command(group)
